Move all type definitions into the Types module.
This commit's a little far-reaching. We've run into a problem where some
new constructors that we need to add to the `LispVal` type contain types
defined in other modules that themselves depend on `LispVal`. This
necessitates cyclic imports, which GHC handles somewhat poorly, so this
commit just moves *all* type definitions into the `Types` module. Note
that the interpreter's codebase only uses qualified imports, so to avoid
changing all occurrences of, say, `Error.LispError` to
`Types.LispError`, a number of said imports were temporarily aliased to
an import of `SchemeInterpreter.Types`. That'll be addressed in a
follow-up commit.